// src/services/openapi/notification-service/api.ts
export interface NotificationReceiverResponseDto {
    /**
     * 
     * @type {number}
     * @memberof NotificationReceiverResponseDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof NotificationReceiverResponseDto
     */
    divisionCode?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationReceiverResponseDto
     */
    roleCode?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationReceiverResponseDto
     */
    positionCode?: string;
}